Add edit article form

+++ main.py
@@ -23,7 +23,7 @@ def index():
 
     articles = []
     q = Article.query(ancestor=account_key)
-    q = q.order(-Article.date_time_last_edited)
+    q = q.order(-Article.date_time_created)
     articles = q.fetch()
 
     if len(articles) < 10:
@@ -54,9 +54,45 @@ def article_create():
     return redirect(url_for('index'))
 
 
+@app.route('/article/<string:article_key_urlsafe>/edit', methods=['POST'])
+def article_edit(article_key_urlsafe):
+    article_key = ndb.Key(urlsafe=article_key_urlsafe)
+    if not article_key:
+        return 'invalid key'
+
+    account = _main._getCurrentAccount()
+    if not account:
+        return 'not logged in'
+
+    account_key = account.key
+
+    if article_key.parent() != account_key:
+        return "cannot edit another person's post"
+
+    article = article_key.get()
+    article.title = request.form['title']
+    article.body = request.form['body']
+    article.date_time_last_edited = datetime.now()
+    article.put()
+
+    return redirect(url_for('index'))
+
+
 @app.route('/article/<string:article_key_urlsafe>/delete', methods=['POST'])
 def article_delete(article_key_urlsafe):
     article_key = ndb.Key(urlsafe=article_key_urlsafe)
+    if not article_key:
+        return 'invalid key'
+
+    account = _main._getCurrentAccount()
+    if not account:
+        return 'not logged in'
+
+    account_key = account.key
+
+    if article_key.parent() != account_key:
+        return "cannot delete another person's post"
+
     if article_key:
         article_key.delete()
     return redirect(url_for('index'))
